add creating Bucket

+++ src/main/scala/ru/pkuznetsov/core/model/Ingredient.scala
@@ -3,7 +3,7 @@ package ru.pkuznetsov.core.model
 import io.circe.generic.semiauto._
 import io.circe.{Decoder, Encoder}
 
-final case class Ingredient(id: Long, name: String, amount: Double, unit: Option[String])
+final case class Ingredient(id: Int, name: String, amount: Double, unit: Option[String])
 
 object Ingredient {
   implicit val decoder: Decoder[Ingredient] = deriveDecoder[Ingredient]
